+++ src/Backend/spItemsRepository.ts
@@ -18,6 +18,7 @@ import {
   newDocumentCallback,
 } from "./spFileHandlers";
 import { getAllItems, getItemFromPosition } from "./spItemsGetters";
+import { refreshDiagnostics } from "../Providers/spLinter";
 
 export class ItemsRepository implements Disposable {
   public fileItems: Map<string, FileItems>;
@@ -37,10 +38,12 @@ export class ItemsRepository implements Disposable {
   }
 
   public handleDocumentChange(event: TextDocumentChangeEvent) {
+    refreshDiagnostics(event.document);
     handleDocumentChange(this, event);
   }
 
   public handleNewDocument(document: TextDocument) {
+    refreshDiagnostics(document);
     newDocumentCallback(this, document.uri);
   }