+++ src/Providers/Linter/registerSPLinter.ts
@@ -1,0 +1,27 @@
+﻿import {
+  workspace as Workspace,
+  window,
+  languages,
+  ExtensionContext,
+} from "vscode";
+
+import { refreshDiagnostics } from "../spLinter";
+import { compilerDiagnostics } from "./compilerDiagnostics";
+import { throttles } from "./throttles";
+
+export function registerSPLinter(context: ExtensionContext) {
+  context.subscriptions.push(languages.createDiagnosticCollection("compiler"));
+  context.subscriptions.push(
+    window.onDidChangeActiveTextEditor((editor) => {
+      if (editor) {
+        refreshDiagnostics(editor.document);
+      }
+    })
+  );
+  context.subscriptions.push(
+    Workspace.onDidCloseTextDocument((document) => {
+      compilerDiagnostics.delete(document.uri);
+      delete throttles[document.uri.path];
+    })
+  );
+}