+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Release Notes
 
+## [3.0.6]
+
+### Fixed
+
+- Fixed a typo in the parser.
+- Switched to output window instead of terminal for the Compile SM command.
+- Fixed malformed code outline (see #179).
+
 ## [3.0.5]
 
 ### Fixed